api/agent_management/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
esprima>=4.0.1
jsonschema>=3.2.0 